requirements.txt
numpy<1.24
lap==0.4.0
scipy
opencv-python <= 4.6.0
pycocotools
PyYAML
tqdm
scikit-learn
Pillow  
matplotlib==3.7.1
scikit-image

# for uad
pandas

